fixes warnings reported by clang or clazy (from QtCreator)
Summary: fixes warnings reported by clang or clazy (from QtCreator)

Test Plan: the automatic tests are OK

Reviewers: #elisa, astippich

Reviewed By: #elisa, astippich

Differential Revision: https://phabricator.kde.org/D12460

+++ src/databaseinterface.h
@@ -192,7 +192,7 @@ private:
     int computeTrackPriority(qulonglong trackId, const QUrl &fileName);
 
     qulonglong internalInsertTrack(const MusicAudioTrack &oneModifiedTrack, const QHash<QString, QUrl> &covers,
-                                   int originTrackId, QSet<qulonglong> &modifiedAlbumIds, TrackFileInsertType insertType,
+                                   qulonglong originTrackId, QSet<qulonglong> &modifiedAlbumIds, TrackFileInsertType insertType,
                                    QList<qulonglong> &newAlbumIds);
 
     MusicAudioTrack buildTrackFromDatabaseRecord(const QSqlRecord &trackRecord) const;
